ai-service/requirements.txt
# AI Service Requirements
requests>=2.32.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.0
httpx>=0.25.0
aiohttp==3.9.3
